moved code from app.component to dashboard.component to enable routing
