+++ iaq-ui/src/app/app.component.html
@@ -1,17 +1,1 @@
-<app-navbar></app-navbar>
-
-<div class="graph-container">
-    <div class="graph-item">
-        <ng-template #viewContainerRef></ng-template>
-    </div>
-</div>
-
-<div class="add-graph-button">
-    <button mat-fab color="primary" (click)="addSensorChart()">
-        <mat-icon>add</mat-icon>
-    </button>
-</div>
-
-<!--
 <router-outlet></router-outlet>
--->
