+++ iaq-ui/src/app/core/core.module.ts
@@ -1,20 +1,17 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { NavbarComponent } from './navbar/navbar.component';
 import { SharedModule } from '../shared/shared.module';
 
 
 
 @NgModule({
   declarations: [
-    NavbarComponent
   ],
   imports: [
     CommonModule,
     SharedModule
   ],
   exports: [
-    NavbarComponent,
   ]
 })
 export class CoreModule { }
