+++ iaq-ui/src/app/feature/dashboard/dashboard.component.html
@@ -1,0 +1,13 @@
+<app-navbar></app-navbar>
+
+<div class="graph-container">
+    <div class="graph-item">
+        <ng-template #viewContainerRef></ng-template>
+    </div>
+</div>
+
+<div class="add-graph-button">
+    <button mat-fab color="primary" (click)="addSensorChart()">
+        <mat-icon>add</mat-icon>
+    </button>
+</div>
