+++ iaq-ui/src/app/feature/dashboard/dashboard.component.sass
@@ -1,0 +1,14 @@
+/*.graph-container
+    display: grid
+    grid-template-columns: repeat(auto-fill, 1fr)
+    grid-template-rows: 1fr
+    grid-column-gap: 0px
+    grid-row-gap: 0px
+
+/*.graph-item
+    padding: 7em
+
+.add-graph-button
+    position: fixed
+    right:1em
+    bottom:1em
