+++ iaq-ui/src/app/feature/dashboard/dashboard.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { DashboardComponent } from './dashboard.component';
+import { SharedModule } from 'src/app/shared/shared.module';
+import { NavbarComponent } from './navbar/navbar.component';
+
+
+
+@NgModule({
+  declarations: [
+    DashboardComponent, 
+    NavbarComponent
+  ],
+  imports: [
+    CommonModule,
+    SharedModule
+  ],
+  exports: [
+    NavbarComponent
+  ]
+})
+export class DashboardModule { }
