+++ iaq-ui/src/app/feature/dashboard/navbar/navbar.component.sass
