+++ iaq-ui/src/app/feature/feature.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ChartDisplayModule } from './chart-display/chart-display.module';
+import { DashboardModule } from './dashboard/dashboard.module';
 
 
 
@@ -9,6 +10,7 @@ import { ChartDisplayModule } from './chart-display/chart-display.module';
   imports: [
     CommonModule,
     ChartDisplayModule,
+    DashboardModule
   ]
 })
 export class FeatureModule { }
